requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.61
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.3
Django==4.2.5
django-address==0.2.8
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.8.0
urllib3==1.26.17
wcwidth==0.1.9
